UXD-837 chore(ListBox): restructure tests
