+++ packages/ListBox/test/specs/single/ListBox.single.spec.js
@@ -1,7 +1,6 @@
 import React from "react";
-import { render, fireEvent, configure, waitFor } from "@testing-library/react";
+import { render, fireEvent, configure } from "@testing-library/react";
 import { Controlled } from "../../../stories/examples/Single/Controlled";
-import { OnChange } from "../../../stories/examples/Single/OnChange";
 import ListBox from "../../../src";
 
 configure({ testIdAttribute: "data-pka-anchor" });
@@ -64,32 +63,6 @@ describe("ListBox single select", () => {
     expect(getByTestId("clear-button")).toBeVisible();
   });
 
-  it("should clear selected option when x is clicked", () => {
-    const { queryByTestId, getByTestId, openSelect, selectVenus } = renderComponent();
-
-    openSelect();
-    selectVenus();
-    fireEvent.click(getByTestId("clear-button"));
-    expect(queryByTestId("clear-button")).toBeNull();
-    expect(getByTestId("list-box-trigger")).not.toHaveTextContent(/venus/i);
-    expect(getByTestId("list-box-trigger")).toHaveTextContent(/select/i);
-  });
-
-  it("should have a filter in dropdown", () => {
-    const { getByText, getByTestId } = render(
-      <ListBox>
-        <ListBox.Filter />
-        <ListBox.Option>Venus</ListBox.Option>
-        <ListBox.Option>Jupiter</ListBox.Option>
-      </ListBox>
-    );
-
-    fireEvent.click(getByText(/Select/));
-    waitFor(() => {
-      expect(getByTestId("list-filter")).toBeInTheDocument();
-    });
-  });
-
   it("should have custom height of 500", () => {
     const { getByTestId } = renderComponent({
       height: 500,
@@ -98,7 +71,6 @@ describe("ListBox single select", () => {
     expect(getByTestId("styled-list").getAttribute("height")).toMatch("500");
   });
 
-  //
   it("should be disabled", () => {
     const { openSelect, popoverIsHidden } = renderComponent({
       isDisabled: true,
@@ -106,166 +78,6 @@ describe("ListBox single select", () => {
 
     openSelect();
     popoverIsHidden();
-  });
-
-  it("should be displayed inline with no Popover", () => {
-    const { queryByTestId } = renderComponent({
-      isInline: true,
-    });
-
-    expect(queryByTestId("popover.content")).toBeNull();
-  });
-
-  it("should focus on option container as soon as the Popover is open", done => {
-    const { openSelect, getByTestId } = renderComponent();
-
-    openSelect();
-    setTimeout(() => {
-      expect(document.activeElement).toBe(getByTestId("popover.content"));
-      done();
-    }, 350);
-  });
-
-  it("should not focus on option container as soon as the Popover is open", done => {
-    const { getByText, getByTestId } = render(
-      <ListBox>
-        <ListBox.Popover key="Popover" shouldKeepFocus />
-        <ListBox.Option>Venus</ListBox.Option>
-        <ListBox.Option>Jupiter</ListBox.Option>
-      </ListBox>
-    );
-
-    fireEvent.click(getByText(/Select/));
-
-    // NOT BEST PRACTICE: fix later, this is due the timer I had to put on the ListBox becaue the PopoverTimer
-    setTimeout(() => {
-      expect(document.activeElement).not.toBe(getByTestId("popover.content"));
-      done();
-    }, 350);
-  });
-
-  it("calls renderTrigger and changes the render method for label", () => {
-    const togglePopover = (dispatch, types) => () => {
-      dispatch({ type: types.togglePopover });
-    };
-
-    const onRenderTrigger = jest.fn((selected, options, { dispatch, propsForTrigger, types, refTrigger }) => {
-      return (
-        <button type="button" {...propsForTrigger()} onClick={togglePopover(dispatch, types)} ref={refTrigger}>
-          Toggle ListBox
-        </button>
-      );
-    });
-
-    const { getByText } = renderComponent({}, [
-      <ListBox.Trigger key="trigger">{onRenderTrigger}</ListBox.Trigger>,
-      [...childrenContent],
-    ]);
-
-    expect(onRenderTrigger).toHaveBeenCalled();
-    expect(getByText(/toggle listbox/i)).toBeInTheDocument();
-    fireEvent.click(getByText(/toggle listbox/i));
-    expect(getByText(/venus/i)).toBeInTheDocument();
-    expect(getByText(/jupiter/i)).toBeInTheDocument();
-  });
-
-  it("should display message when filter input does not find a match", () => {
-    const { getByTestId, getByText } = render(
-      <ListBox isMulti>
-        <ListBox.Filter noResultsMessage="No match" />
-        <ListBox.Option>Venus</ListBox.Option>
-        <ListBox.Option>Jupiter</ListBox.Option>
-      </ListBox>
-    );
-
-    fireEvent.click(getByText(/Select/));
-    fireEvent.change(getByTestId("list-filter-input"), { target: { value: "g" } });
-    waitFor(() => {
-      expect(getByTestId("no-results")).toBeInTheDocument();
-      expect(getByText("No match")).toBeInTheDocument();
-    });
-  });
-
-  it("placeholder should display default label when no option is selected", () => {
-    const { getByText } = renderComponent();
-
-    expect(getByText(/Select/i)).toBeInTheDocument();
-  });
-
-  it("placeholder should display custom label when no option is selected", () => {
-    const { getByText } = renderComponent({
-      placeholder: "Choose an option!",
-    });
-
-    expect(getByText("Choose an option!")).toBeInTheDocument();
-  });
-
-  it("should have popover open already ", () => {
-    const { popoverIsVisible } = renderComponent({
-      isOpen: true,
-    });
-
-    popoverIsVisible();
-  });
-
-  it("onChange should have a correct signature (selectedIndex, options, more)", () => {
-    const customChildrenContent = [
-      <ListBox.Option key="1" value="11">
-        Venus
-      </ListBox.Option>,
-      <ListBox.Option key="2" value="21">
-        Jupiter
-      </ListBox.Option>,
-    ];
-
-    function onChange(selectedIndex, options, more) {
-      expect(arguments.length).toBe(3);
-      expect(typeof selectedIndex).toBe("number");
-
-      expect(options[0].value).toBe("11");
-      expect(options[1].value).toBe("21");
-      expect(options[0].label).toBe("Venus");
-      expect(options[1].label).toBe("Jupiter");
-
-      expect("actionTypes" in more).toBeTruthy();
-      expect("eventType" in more).toBeTruthy();
-    }
-
-    const { openSelect, selectVenus } = renderComponent(
-      {
-        onChange,
-      },
-      [...customChildrenContent]
-    );
-
-    openSelect();
-    selectVenus();
-  });
-
-  // onClickClear passes even when clear button is not shown on the UI after selecting an option from the popover.
-  // Because of CSS the clear button is still present
-  it("calls onClickClear event when clicking clear button", () => {
-    const onClickClearTrigger = jest.fn();
-    const { getByTestId, openSelect, selectVenus } = renderComponent({}, [
-      <ListBox.Trigger key="trigger" onClickClear={onClickClearTrigger} />,
-      [...childrenContent],
-    ]);
-
-    openSelect();
-    selectVenus();
-    fireEvent.click(getByTestId("clear-button"));
-    expect(onClickClearTrigger).toHaveBeenCalled();
-  });
-
-  it("should not render the 'x' clear button", () => {
-    const { queryByTestId, openSelect, selectVenus } = renderComponent({}, [
-      <ListBox.Trigger key="trigger" hasClearButton={false} />,
-      [...childrenContent],
-    ]);
-
-    openSelect();
-    selectVenus();
-    expect(queryByTestId("clear-button")).toBeNull();
   });
 
   it("should select an option via a controlled button", () => {
@@ -293,25 +105,5 @@ describe("ListBox single select", () => {
 
     expect(getAllByTestId("list-option--is-selected").length).toBe(1);
     expect(getByTestId("list-option--is-selected").textContent).toBe("Spiderman");
-  });
-
-  it("should not create a stale state when reading the state on handleChange", () => {
-    const log = [];
-    const storeLog = msg => log.push(msg);
-    const originalConsoleLog = console.log;
-    console.log = jest.fn(storeLog);
-
-    const { getByText } = render(<OnChange />);
-
-    const options = [getByText(/Wonder Woman/), getByText(/Thor/), getByText(/Batman/)];
-    fireEvent.click(options[0]);
-    fireEvent.click(options[1]);
-    fireEvent.click(options[2]);
-
-    expect(log.includes(null)).toBe(true);
-    expect(log.includes("Wonder Woman")).toBe(true);
-    expect(log.includes("Thor")).toBe(true);
-    expect(log.includes("Batman")).toBe(false);
-    console.log = originalConsoleLog;
   });
 });
